d1send_m1pro/requirements.txt
google-api-python-client
google-api-core>=2.29.0
google-auth
google-auth-oauthlib
